Fri Aug 20 19:13:30 PDT 1999 (Raster)
Fixed some iconification bugs - a bit cleaner now - wiht snapshots and
everything. stil soem annoying niggly where if you are snapshotting the ewin
and its shaded when you iconify - de-iconifying leaves the client window
exposed but the border is still shaded - E seems to be responding to an
events after de-iconification - mayeb - but what it is - i dunno.


git-svn-id: http://svn.enlightenment.org/svn/e@68 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33