Tue Oct 26 13:42:53 PDT 1999 (KainX)
First off, I removed some unneeded variables from E-MemWatch.  I also fixed a
reporting bug in E-Disk.

But most importantly, I fixed Epplet_change_label() to avoid changing a label
to the exact same string.  This is the best place to handle this situation,
because it saves resources (we already have both strings to compare) and, due
to the nature of epplets, the probability is very high that this function
would only (or at least primarily) be called from a timer callback.  So it
simplifies client code this way too.


git-svn-id: http://svn.enlightenment.org/svn/e@983 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33