Wed Dec  1 19:11:13 PST 1999 (KainX)
For one thing, I think I've fixed the seg fault in E-MoonClock.  Here's
hoping.

I also added the background setting functionality to E-Slides, but this time
it was done *right*. :-)  Take a look and see for yourself.  If you look at
the code, it also demonstrates how to use E IPC to set backgrounds.


git-svn-id: http://svn.enlightenment.org/svn/e@1476 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33