Fri Mar 17 23:01:24 PST 2000 (KainX)
I found the problem in icccm.c.  For some reason, somebody thought it
would be a good idea to set minimum width/height equal to the base
width/height if no *maximum* size was specified.  I think they meant to
check for *minimum* size.  Anyway, I fixed that and added a more robust
check for that kind of thing.


git-svn-id: http://svn.enlightenment.org/svn/e@2271 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33