Urm. oops. hehe :)
Reversed that change, as it caused segfaults in free().

I can't grok this code, so I'll leave it to raster.

Basically, if my code calls imlib_free_image_and_decache(), it doesn't free
the image, and it leaks.

The fix I applied made sure _Imlib_ConsumeImage was called, but this caused
a segfault.

*baffle


git-svn-id: http://svn.enlightenment.org/svn/e@3239 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33