Okay. Efsd's filetype module can now use values it reads from a file dynamically in the file type. This is useful for things like version numbers -- so far I had to use things like
>9      byte            49      1.x
>9      byte            50      2.x
>9	byte		51	3.x

now I can simply do

>9      byte            .       %d.x

(a dot is my way of saying "value doesn't matter"). There are cases
where this isn't too meaningful for file types, such as a hypothetical
image/jpg/352x467 because noone will ever match that precise value.
The functionality is there though, and I'll make good use of it in
Efsd's next feature, file descriptions (basically equivalents of the
file(1) output). e17 could then show a popup or something that contains
the received description when the user leaves the pointer over an icon
for a while etc.


git-svn-id: http://svn.enlightenment.org/svn/e@5154 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33