Okay, the xml backend for the filetype definitions is in. Turns out that the XML parsing is easily fast enough, and a *lot* faster than the edb approach previously used. Also, the XML file has only about 2.6 times the size of the original magic.txt, and that is without any compression and with linebreaks and indentation. It is probably a good idea to make uninstall before updating or manually remove the db files in /usr/local/share/efsd, or wherever they get installed on your machine.
I use a SAX parser for reading the configs, which should keep
memory requirements minimal. I do need to build a full
document tree though when the filetype definitions are written
out to files, but that is temporary, hardly noticeable anyway
and only happens when requested.

It was a good opportunity to clean up a good deal of things,
for example the separation of file magic and pattern
definition files is gone, because they're both stored in the
same file (see tools/filetypes.dtd for details), also users
can specify their own magic tests and not only pattern tests
now. I've also fixed a few bugs in the cleanup handling on
exit and the reloading behaviour when the XML files are
modified while efsd is running.

Raster's MRU heuristic suggestion is in, too. I think the
difference is noticeable :)

There's one catch that I totally missed in our discussion -- I
need Edb anyway. Oops. Of course the entire metadata handling
is edb based, and edb is certainly a good solution here. So
I've added an additional dependency. Personally I can live
with that, if someone has problems with that, flame away, but
I'm very happy with the way XML helps me in handling the case
at hand.


git-svn-id: http://svn.enlightenment.org/svn/e@5257 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33