Here's libefsd command queue handling. When Efsd cannot send all commands as necessary, commands are now queued on the client side. The queueing is transparent to the client, i.e. when commands need to be queued, Efsd calls still return the command ID, as usual. You can check whether commands are queued at any time using efsd_commands_pending(), and efsd_flush() to flush the queue. This call will *not* block. It tries to write as many queued commands as possible, but always returns. The idea here is to never see a client app hanging in case Efsd screws up. efsd_flush() returns a value > 0 when the queue could be flushed completely.
The queue is reduced as much as possible with every Efsd command sent.

Raster, as far as I can see the correct way to handle this would be to modify Ecore to also
allow event callbacks to be executed when an fd becomes ready to be *written to*. When
efsd_commands_pending() tells you that commands are queued, add an fd event handler when the
Efsd connection fd becomes writeable, and then call efsd_flush() in the handler. When
efsd_flush()s return value indicates more commands are in the queue, keep setting the handler.

I have not modified anything in e17 yet. Here's what I currently see when I set all the
metadata on a large directory: for a while all works well, and depending on the timing of
things, the queue first builds up and then gets eliminated or keeps increasing until all
metadata commands are issued. Some are still hanging in the queue at this time, because there's
no logic yet to smartly flush the queue. You'll then notice that the queue is worked up as soon
as you issue any other efsd commands, e.g. when opening another view.

Btw I don't see why it's necessary to set all that x/y metadata to 9999999 when a view is first
opened? I think it would make more sense to make the icon layout smarter and use default
layouts with overriding x/y coordinates when they're available.


git-svn-id: http://svn.enlightenment.org/svn/e@5404 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33