E_EVENT system is now a bit map, allowing one observer to watch MANY events, for instance: E_EVENT_BORDER_ICONIFY | E_EVENT_BORDER_UNICONIFY | E_EVENT_BORDER_NEW, etc. I also made it so that the current event is passed when notifying, so that if watching multiple events, you can do different things for each.
added an e_observer_notify_all_observees() to notify all existing observees of things, such as the creation of a border. This way, an observer can watch for E_EVENT_BORDER_NEW events and register all new borders.

Iconification! Click that iconify button. Then click the middle button on the desktop to get a menu of iconified apps to uniconify. This menu is a decent example of using the observer code to avoid polling or rebuilding more than necessary. Next, maybe instead of completely rebuilding each time something changes, maybe it should dynamically add or remove the respective menu items?


git-svn-id: http://svn.enlightenment.org/svn/e@5854 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33