This isn't 100% right yet I'm pretty sure, any input appreciated.
Updated init.d/elogin script - it works from a command line, but having it
run on boot leads to an X session with no keyboard input.  I've spent a bit
of time in single user mode today, so be warned =)

The nature of the spawner is a little different, it forks and exits so you
can get back to a command line after launching.

It now has a pidfile and responds to HUPs

Should prolly add elogind -d $DISPLAY shortly.


git-svn-id: http://svn.enlightenment.org/svn/e@6281 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33