Added turntable option so you can spin K's head around.
Fire up entice with the following image and play with the 't' key.
Takes a little while to start as the images are pre-rendered --
this should probably be fixed. Hit enter, change image or perform
an edit opperation to stop the turntable.

http://www.vergenet.net/~horms/tmp/k.jpg

Be careful with large images. This option can chew up a lot
of memory.


git-svn-id: http://svn.enlightenment.org/svn/e@6692 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33