put e_container_scroll_set/get() in the header file. these can be used to do custom scrolling
also, adjust the scroll_offset on resize to keep you in the same position relative to the elements. if the length of all the elements becomes smaller than the size of the container, it will reset the offset to 0 to prevent scrolling from getting frozen out.

do the same thing on element remove. (i haven't tested this yet, not sure exactly how it will work)


git-svn-id: http://svn.enlightenment.org/svn/e@7693 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33