Fri Oct 31 15:21:53 2003                        Michael Jennings (mej)
Fixes to string class find/splice routines.

Added SPIF_DEFINE_PROPERTY_FUNC_NONOBJ() macro for defining get/set
methods of a property which is not an object.  It should be used in
place of SPIF_DEFINE_PROPERTY_FUNC() for spif_*_t types that are not
objects, such as spif_int32_t and spif_sockport_t.


git-svn-id: http://svn.enlightenment.org/svn/e@7774 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33