Twins can now have 'extra' gevasobj attached which when event handlers are setup correctly will allow the main/aux object to be dragged around by the 'extra' objects attached to the twin.
This is quite handy for loading edje objects, labeling them (the
twin's aux object) and throwing a big object over the lot (an
'extra' object) so that the extra object drags the main/aux
around with it. Set the extra object to cover main/aux and be at
a level above them.


git-svn-id: http://svn.enlightenment.org/svn/e@12231 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33