Added context switching to Engage. This means that you can have something like: ~/.e/e/applications/engage/.Applications/.order ~/.e/e/applications/engage/.Backgrounds/.order ~/.e/e/applications/engage/.SomethingElse/.order
and when you right click engage, you will see a menu that allows you to
switch between contexts. You can see a sample here:

http://www.uk.get-e.org/Screenshots/User_Submitted_previews/2005.04.26.desky.jpg.html

This turns Engage into an E17 backgrounds selection tool.
The following limitations apply:
1) dir names need to start with a "."
2) an initial ~/.e/e/applications/engage/.order file needs to be there

If an icon list is too large for the screen, it will not scroll at the
moment. This is to be fixed later. Currently, also, there is no way to filter
out the appearence of icons for running apps in a certain state. This might
be fixed at some stage.


git-svn-id: http://svn.enlightenment.org/svn/e@14387 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33