* Add support of yuy2 to emotion_xine module. So now, we can play wmv movies!
I had to make some module API changes:
 - rename yuv_size_get to video_data_size_get
 - add the function "Emotion_Format (*format_get) (void *ef);" in Emotion_Video_Module. This function should return the format used by the video (EMOTION_YV12, EMOTION_YUY2 or EMOTION_BGRA).

I use a yuy2->bgra converter since evas doesn't support YUY2. This converter should be rewrite since it's really not optimized and since it is under GPLv2 (emotion is under BSD). I added an explicit comment above the function for that, but it'd be better to use a converter under BSD.
If it's really incompatible with emotion license, I will remove it.


git-svn-id: http://svn.enlightenment.org/svn/e@15675 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33