Initial checkin of entrance_edit_gui into CVS.
entrance_edit_gui (the entrance config panel), is currently being built in a funky way. With Codewarrior's help we came up with a basic idea in etk. I've gone ahead to refactor out (if you will), an easier and higher level api ontop of what we have. Basically, the api is focusing on converging a _very_ limited subset of etk (initially), to allow quickly putting together the entrance config panel. I'm not under any illusions of grandoise plans for this api, its strictly my personal bitch in the entrance gui config building industry, and by that definition its exposed api is supposed to be simple (even naively so). With LinuxTitan's help, we hope to spread the thin layer over ewl eventually.

I've checking this into proto for a couple of reasons:

	- The whole code is going to be in a constant state of flux for the next couple of days/weeks. I don't want to move it to entrance in this state of flux, currently, its leaking everywhere... no frees, no error checks, nothing, its just a piece of scape code (or goat?), that will be eventually beaten into shape. Once things naturally slow down, I'll make that move.

	- xcomp doesn't want entrance to have a hard dependency on any toolkit. So the only config mechanism that will _always_ be there is entrance_edit (commandline). Once you have the neccessary toolkit installed, then the gui can also be built. I assume most ppl will want this anyhoo. To help us to keep things very loosely coupled, we'll be workign out of proto till things are a bit more stable.
	- This design is not final. Things can still change _drastically_.

ok... i need to sleep... :)


git-svn-id: http://svn.enlightenment.org/svn/e@24580 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33