not complete yet - actually i need help with the fragment shader.
in evas_gl_texture.c i have a frag shader, and it tries to use a set of 3
textures that act as the yuv planes, BUT the u and v textures (Utex and Vtex)
are simply getting values from the Ytex - regardless of what i try. grrr.
what's up with that?


git-svn-id: http://svn.enlightenment.org/svn/e@27495 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33