Line, Polygon and minor fixes for software_16.
Line is a complete rewrite based on my university works. It's much
cleaner than the engine/common and works better (the later is
producing weird results, I still have to debug why), but I don't
provide anti-aliased drawings.

Polygon is almost the same code, with minor changes to draw the spans
as soon as possible and then no malloc/free is required for each of
them.

Minor fixes to remove unused variables, gotos...


git-svn-id: http://svn.enlightenment.org/svn/e@32161 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33