Add Red Black tree API. Right now only insertion, lookup and iterator work.
The current implementation choose to move the node allocation outside of eina
control like eina_inlist. They currently have the same memory footprint as
eina_inlist and the implementation of insertion and lookup are iterative
making it quite fast. This should make them a good competitor of eina_inlist
for eina_hash and eina_stringshare.



git-svn-id: http://svn.enlightenment.org/svn/e@35689 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33