Recommit Joerg Sonnenberger patch with more fix.
When writing jpeg scanline, it wouldn't fail with FILE* as an output, but will
with a data buffer. So we need to re run the jpeg routine on this scanline. As I
don't know if the buffer is always preserved during two calls to the jpeg
library, I am copying it again.



git-svn-id: http://svn.enlightenment.org/svn/e@35804 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33