e_dbus: signals are per-connection, fix signal match removal.
Each connection must have its own signal handler lists. This patch
also fixes match removal, it was using the sh->PIECE instead of PIECE
to give it names, it's also better split the buffer manipulation so it
checks for overflows as well.

NOTE: signal handling is still unoptimal, it should be handled per
connection instead of a global ecore signal.



git-svn-id: http://svn.enlightenment.org/svn/e@36639 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33