eina_bool is now an unsigned char.
This makes clear what size and if we have a signal or not in our type,
so no compilers have the right to choose different for us (like they
do with enum).

Also fixes compile of C++ programs with eina_rectangle.h, that
contains some inline code, the expression evaluates a boolean
expression and compiler does not know how to convert that to old
Eina_Bool.



git-svn-id: http://svn.enlightenment.org/svn/e@36939 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33