enlightenment_fm_open to open directories from command line.
This command line tool will make it easier to open directories from
command line, instead of calling
   enlightenment_remote -exec-action fileman "PATH"

It has the advantage of being much simpler and smaller and also
resolving the realpath for you.

If there is no such action "fileman" (ie: module is not loaded), then
it will return failure. To implement this I had to change the
-exec-action so it returns a boolean (it will report that value with
enlightenment_remote as well).




git-svn-id: http://svn.enlightenment.org/svn/e@37861 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33